Flask/requirements.txt
Flask==2.3.3
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.76
Werkzeug==2.3.7